fix(about): resolve globe visualization hydration and add error handling
- Add ErrorBoundary component for graceful failure handling
- Move Globe component to separate file for better code organization
- Fix hydration mismatch by moving random arc generation to client-side
- Add proper TypeScript interfaces for arc data
- Implement loading fallback with pulse animation
- Add dynamic import with SSR disabled for Globe component
- Handle mounted state to prevent server/client rendering mismatch
- Add error logging for debugging globe visualization issues
- Improve code modularity with separate component files
