+++ src/components/about/about-section.tsx
@@ -2,20 +2,9 @@
 
 import { motion, useInView } from 'framer-motion';
 import { Globe as GlobeIcon, Users, Zap } from 'lucide-react';
-import { useTheme } from 'next-themes';
-// eslint-disable-next-line import/order
-import dynamic from 'next/dynamic';
-import { Suspense, useEffect, useRef, useState } from 'react';
+import { useRef } from 'react';
 
-// Dynamically import Globe with no SSR
-const Globe = dynamic(() => import('react-globe.gl'), {
-  ssr: false,
-  loading: () => (
-    <div className="flex h-[400px] w-[400px] items-center justify-center sm:h-[600px] sm:w-[600px]">
-      <div className="h-32 w-32 animate-pulse rounded-full bg-primary/10" />
-    </div>
-  ),
-});
+import { GlobeVisualization } from './globe';
 
 const stats = [
   {
@@ -37,59 +26,6 @@ const stats = [
     description: 'Round-the-clock assistance',
   },
 ];
-
-// Pre-calculate arcs data
-const N = 15;
-const arcsData = [...Array(N)].map(() => ({
-  startLat: (Math.random() - 0.5) * 180,
-  startLng: (Math.random() - 0.5) * 360,
-  endLat: (Math.random() - 0.5) * 180,
-  endLng: (Math.random() - 0.5) * 360,
-  color: ['#FF6B6B', '#4ECDC4', '#45B7D1'][Math.round(Math.random() * 2)],
-}));
-
-function GlobeVisualization() {
-  const globeRef = useRef();
-  const { theme } = useTheme();
-  const [dimensions, setDimensions] = useState({ width: 600, height: 600 });
-
-  useEffect(() => {
-    const updateDimensions = () => {
-      const width = window.innerWidth < 640 ? 400 : 600;
-      setDimensions({ width, height: width });
-    };
-
-    updateDimensions();
-    window.addEventListener('resize', updateDimensions);
-    return () => window.removeEventListener('resize', updateDimensions);
-  }, []);
-
-  const globeImageUrl =
-    theme === 'dark'
-      ? '//unpkg.com/three-globe/example/img/earth-dark.jpg'
-      : '//unpkg.com/three-globe/example/img/earth-blue-marble.jpg';
-
-  return (
-    <div className="relative">
-      <Globe
-        ref={globeRef}
-        width={dimensions.width}
-        height={dimensions.height}
-        globeImageUrl={globeImageUrl}
-        backgroundColor="rgba(0,0,0,0)"
-        arcsData={arcsData}
-        arcColor="color"
-        arcDashLength={0.4}
-        arcDashGap={4}
-        arcDashAnimateTime={1500}
-        arcStroke={1}
-        atmosphereColor="#007AFF"
-        atmosphereAltitude={0.1}
-        enablePointerInteraction={false}
-      />
-    </div>
-  );
-}
 
 export function AboutSection() {
   const containerRef = useRef(null);
@@ -156,15 +92,7 @@ export function AboutSection() {
             transition={{ duration: 0.7 }}
             className="relative flex items-center justify-center lg:justify-end"
           >
-            <Suspense
-              fallback={
-                <div className="flex h-[400px] w-[400px] items-center justify-center sm:h-[600px] sm:w-[600px]">
-                  <div className="h-32 w-32 animate-pulse rounded-full bg-primary/10" />
-                </div>
-              }
-            >
-              <GlobeVisualization />
-            </Suspense>
+            <GlobeVisualization />
           </motion.div>
         </div>
       </div>
